Moved placeholder blog card to an appropriate file
